requirements.txt
langchain
jupyterlab
notebook
python-dotenv
openai
pandas
tabulate
wikipedia
duckduckgo-search